proteccion errores

+++ ft_printf.h
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: amorion- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/08/17 09:01:29 by amorion-          #+#    #+#             */
-/*   Updated: 2021/08/17 10:26:22 by amorion-         ###   ########.fr       */
+/*   Created: 2021/08/18 10:59:09 by amorion-          #+#    #+#             */
+/*   Updated: 2021/08/18 10:59:45 by amorion-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,5 @@ int		ft_print_arg(va_list ap, char c);
 int		ft_print_address(unsigned long long p);
 void	ft_putnbr_fd(int n, int fd);
 void	ft_putstr_fd(char *str, int fd);
-size_t	ft_strlen(char *str);
+int		ft_strlen(char *str);
 #endif
